README
This website is based for a model Car Rental company which provides self drive cars and Cars with Drivers. 
The user is charged based on the duration he/she hires the car. Cars with drivers are charged per day basis. 

The site contains a separate user interface and an admin interface.